Did the item stuff

+++ server/world/level/item/container_loot.go
@@ -2,5 +2,5 @@ package item
 
 type ContainerLoot struct {
 	LootTable string `nbt:"loot_table"`
-	Seed      int32  `nbt:"seed"`
+	Seed      int64  `nbt:"seed"`
 }
